Create IDCardGenerator.java
When the user selects the "Generate New ID Card" option, we prompt the user to input their personal information and generate an ID card with the entered information. We then store the ID card information in the hashmap using the user's ID number as the key. When the user selects the "View Existing ID Cards" option, we display all of the stored ID cards by iterating over the hashmap.